requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
PyMuPDF
python-docx
langchain
langchain-google-genai
faiss-cpu
